pr amends: update DeckTitle file structure, exclude __mocks__ from build output, bump package version
